Responded to review and improved decoderawtransaction output
